created a src folder and move all folders inside the src folder - Basic setup done
