+++ backend/src/response/lng/en.json
@@ -4,5 +4,6 @@
     "SIGNUP_SUCCESSFULL": "User registered successfully",
     "LOGIN_SUCCESSFULL": "Logged-in successfully",
     "INTERNAL_SERVER_ERROR": "Internal server error",
-    "INVALID_ROUTE": "Invalid route: The route you are trying to access doesn't exists"
+    "INVALID_ROUTE": "Invalid route: The route you are trying to access doesn't exists",
+    "VALID_ROUTE_BUT_INVALID_METHOD": "The requested HTTP method is not allowed on this route"
 }
